requirements.txt
weaviate-client>=4.0.0
langchain
langchain-community
langchain-huggingface
langchain-weaviate
transformers
sentence-transformers
pandas
python-dotenv